--- extra/linux-03.20.00.13/include/linux/netfilter/xt_MARK.h
@@ -1,10 +1,0 @@
-#ifndef _XT_MARK_H_target
-#define _XT_MARK_H_target
-
-#include <linux/types.h>
-
-struct xt_mark_tginfo2 {
-	__u32 mark, mask;
-};
-
-#endif /*_XT_MARK_H_target */
