+++ extra/linux-03.20.00.13/net/netfilter/xt_MARK.c
@@ -1,9 +1,9 @@
 /*
- *	xt_MARK - Netfilter module to modify the NFMARK field of an skb
+ *	xt_mark - Netfilter module to match NFMARK value
  *
  *	(C) 1999-2001 Marc Boucher <marc@mbsi.ca>
  *	Copyright © CC Computer Consultants GmbH, 2007 - 2008
- *	Jan Engelhardt <jengelh@computergmbh.de>
+ *	Jan Engelhardt <jengelh@medozas.de>
  *
  *	This program is free software; you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License version 2 as
@@ -12,45 +12,42 @@
 
 #include <linux/module.h>
 #include <linux/skbuff.h>
-#include <linux/ip.h>
-#include <net/checksum.h>
 
+#include <linux/netfilter/xt_mark.h>
 #include <linux/netfilter/x_tables.h>
-#include <linux/netfilter/xt_MARK.h>
 
 MODULE_LICENSE("GPL");
 MODULE_AUTHOR("Marc Boucher <marc@mbsi.ca>");
-MODULE_DESCRIPTION("Xtables: packet mark modification");
-MODULE_ALIAS("ipt_MARK");
-MODULE_ALIAS("ip6t_MARK");
+MODULE_DESCRIPTION("Xtables: packet mark match");
+MODULE_ALIAS("ipt_mark");
+MODULE_ALIAS("ip6t_mark");
 
-static unsigned int
-mark_tg(struct sk_buff *skb, const struct xt_target_param *par)
+static bool
+mark_mt(const struct sk_buff *skb, const struct xt_match_param *par)
 {
-	const struct xt_mark_tginfo2 *info = par->targinfo;
+	const struct xt_mark_mtinfo1 *info = par->matchinfo;
 
-	skb->mark = (skb->mark & ~info->mask) ^ info->mark;
-	return XT_CONTINUE;
+	return ((skb->mark & info->mask) == info->mark) ^ info->invert;
 }
 
-static struct xt_target mark_tg_reg __read_mostly = {
-	.name           = "MARK",
-	.revision       = 2,
+static struct xt_match mark_mt_reg __read_mostly = {
+	.name           = "mark",
+	.revision       = 1,
 	.family         = NFPROTO_UNSPEC,
-	.target         = mark_tg,
-	.targetsize     = sizeof(struct xt_mark_tginfo2),
+	.match          = mark_mt,
+	.matchsize      = sizeof(struct xt_mark_mtinfo1),
 	.me             = THIS_MODULE,
 };
 
-static int __init mark_tg_init(void)
+static int __init mark_mt_init(void)
 {
-	return xt_register_target(&mark_tg_reg);
+	return xt_register_match(&mark_mt_reg);
 }
 
-static void __exit mark_tg_exit(void)
+static void __exit mark_mt_exit(void)
 {
-	xt_unregister_target(&mark_tg_reg);
+	xt_unregister_match(&mark_mt_reg);
 }
 
-module_init(mark_tg_init);
-module_exit(mark_tg_exit);
+module_init(mark_mt_init);
+module_exit(mark_mt_exit);
