--- extra/linux-devkit/arm-none-linux-gnueabi/usr/include/qtopia/phonon/phonon
@@ -1,34 +1,0 @@
-#ifndef QT_PHONON_MODULE_H
-#define QT_PHONON_MODULE_H
-#include "abstractaudiooutput.h"
-#include "abstractmediastream.h"
-#include "abstractvideooutput.h"
-#include "addoninterface.h"
-#include "audiooutput.h"
-#include "audiooutputinterface.h"
-#include "backendcapabilities.h"
-#include "backendinterface.h"
-#include "effect.h"
-#include "effectinterface.h"
-#include "effectparameter.h"
-#include "effectwidget.h"
-#include "mediacontroller.h"
-#include "medianode.h"
-#include "mediaobject.h"
-#include "mediaobjectinterface.h"
-#include "mediasource.h"
-#include "objectdescription.h"
-#include "objectdescriptionmodel.h"
-#include "path.h"
-#include "phonondefs.h"
-#include "phononnamespace.h"
-#include "platformplugin.h"
-#include "seekslider.h"
-#include "streaminterface.h"
-#include "videoplayer.h"
-#include "videowidget.h"
-#include "videowidgetinterface.h"
-#include "volumefadereffect.h"
-#include "volumefaderinterface.h"
-#include "volumeslider.h"
-#endif
