--- lms2012/lmssrc/TEST/Makefile
@@ -1,9 +1,0 @@
-APP = TEST
-APP_PATH = prjs
-INSTALL_TO = X86
-
-PROGRAMS = Change+Name Debug Display Leds Motors NoDebug Performance Power \
-           SDCard Sensors Sound UnlockDatalog USlevel
-SOUNDS = Test
-
-include ../rules.mk
