--- lms2012/lmssrc/Test/Makefile
@@ -1,12 +1,0 @@
-APP = Test
-APP_PATH = tools
-
-PROGRAMS = Test
-ICON = TestIcon
-GRAPHICS = pop3 pop6
-SOUNDS = Test
-
-pop3_SOURCE = 144x65_POP3
-pop6_SOURCE = 144x116_POP6
-
-include ../rules.mk
